cleanup and document benchmark
